style(web-demo): organize components into folders
