+++ packages/pixelmanipulator-web-demo/src/components/footer/Footer.vue
