+++ packages/pixelmanipulator-web-demo/src/components/footer/VersionNumbers.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup>
 import { version as pm_version } from 'pixelmanipulator'
-import * as package_json from '../../package.json'
+import * as package_json from '../../../package.json'
 </script>
 <template>
 	Library v{{pm_version}},
