+++ packages/pixelmanipulator-web-demo/src/components/settings/AppSettings.vue
