+++ packages/pixelmanipulator-web-demo/src/components/settings/AppSizes.vue
