+++ packages/pixelmanipulator-web-demo/src/components/settings/FpsAmount.vue
