+++ packages/pixelmanipulator-web-demo/src/components/settings/FpsRadio.vue
