+++ packages/pixelmanipulator-web-demo/src/components/settings/PixelCounterT.vue
