+++ packages/pixelmanipulator-web-demo/src/components/settings/ShFocusBox.vue
