+++ packages/pixelmanipulator-web-demo/src/components/settings/ShTargeter.vue
