+++ packages/pixelmanipulator-web-demo/src/components/settings/ZoomSizes.vue
